backend/requirements.txt
# IncluScore - Dynamic Credit Scoring System
# Python Dependencies for Backend (Compatible with Python 3.13)

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Explainable AI (may need to install from source)
shap>=0.42.0

# API and Validation
pydantic>=2.4.0
python-multipart>=0.0.6

# Database
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# For demo script and testing
requests>=2.28.0
tabulate>=0.9.0